format layout
