+++ app/components/footer/footer.module.css
@@ -1,0 +1,7 @@
+.footer {
+    background-color: aquamarine;
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 160px;
+}
