+++ app/components/footer/index.tsx
@@ -1,4 +1,5 @@
+import styles from "./footer.module.css";
 const Footer = () => {
-    return <div>Footer</div>
-}
-export default Footer
+    return <div className={styles.footer}>Footer</div>;
+};
+export default Footer;
