+++ app/components/header/header.module.css
@@ -1,0 +1,6 @@
+.header {
+    display: flex;
+    justify-content: space-between;
+    margin: 0 32px;
+    background-color: green;
+}
