+++ app/globals.css
@@ -5,3 +5,12 @@
   min-height: 50vh;
   margin: auto;
 }
+
+body {
+  margin: 0 auto;
+  min-height: 100vh;
+}
+
+.main-layout {
+  display: flex;
+}
