+++ app/layout.tsx
@@ -1,6 +1,8 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
+import { Footer, Header, Navbar } from "./components";
+import React from "react";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -17,8 +19,12 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <div>This is a main layout</div>
-        {children}
+        <Header />
+        <div className="main-layout">
+          <Navbar />
+          {children}
+        </div>
+        <Footer />
       </body>
     </html>
   );
