+++ app/page.tsx
@@ -1,14 +1,11 @@
-import Image from "next/image";
-import styles from "./page.module.css";
-import {Header, Footer, Navbar} from "./components";
+// import Image from "next/image";
+// import styles from "./page.module.css";
+// import {Header, Footer, Navbar} from "./components";
 
 export default function Home() {
   return (
     <>
-      <Header/>
-      <Navbar />
       <div>Home page</div>
-      <Footer/>
     </>
   );
 }
